it compiles again

+++ library/modules/runtime/graphics/source/win_d3d12/basebuffer_d3d12.hpp
@@ -19,7 +19,7 @@ namespace tiki
 								BaseBuffer();
 								~BaseBuffer();
 
-		bool					create( GraphicsSystem& graphicsSystem, size_t size, bool dynamic, GraphicsBufferType binding, const void* pInitData = nullptr );
+		bool					create( GraphicsSystem& graphicsSystem, uint size, bool dynamic, GraphicsBufferType binding, const void* pInitData = nullptr );
 		void					dispose( GraphicsSystem& graphicsSystem );
 		
 	private:
